Allow optional authentication for World Shards
The new "cards" section of the NS API can only be queried using worldRequest. A forthcoming shard called "Packs" will be available under the same system, unable to be queried with the nationRequest function but able to be called with the worldRequest function. This shard requires authentication, so what I've done should add authentication to the worldRequest so that I can continue to use this wrapper for the cards section of the API.